% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributesChartsWrapper.R
\name{attributesCharts}
\alias{attributesCharts}
\title{Control Charts for Attributes}
\usage{
attributesCharts(
  data = NULL,
  version = "1",
  attributesChart = "defectives",
  attributesChartDefectivesChartType = "npChart",
  attributesChartDefectsChartType = "cChart",
  defectiveOrDefect = list(types = list(), value = ""),
  plotHeight = 320,
  plotWidth = 480,
  report = FALSE,
  reportAppraiser = TRUE,
  reportAppraiserText = "",
  reportFrequency = TRUE,
  reportFrequencyText = "",
  reportId = TRUE,
  reportIdText = "",
  reportMeasurementName = TRUE,
  reportMeasurementNameText = "",
  reportMeasusrementSystemName = TRUE,
  reportMeasusrementSystemNameText = "",
  reportMetaData = TRUE,
  reportPerformedBy = TRUE,
  reportPerformedByText = "",
  reportSubgroupSize = TRUE,
  reportSubgroupSizeText = "",
  reportTime = TRUE,
  reportTimeText = "",
  reportTitle = TRUE,
  reportTitleText = "",
  timeStamp = list(types = list(), value = ""),
  total = list(types = list(), value = "")
)
}
\description{
Control Charts for Attributes
}
