% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeAnalysisWrapper.R
\name{doeAnalysis}
\alias{doeAnalysis}
\title{Analyse Design}
\usage{
doeAnalysis(
  data = NULL,
  version = "1",
  blocksFactorial = list(types = list(), value = ""),
  blocksResponseSurface = list(types = list(), value = ""),
  codeFactors = TRUE,
  codeFactorsManualTable = list(optionKey = "predictors", types = list(), value = list()),
  codeFactorsMethod = "automatic",
  continuousFactorsFactorial = list(types = list(), value = list()),
  continuousFactorsResponseSurface = list(types = list(), value = list()),
  contourSurfacePlot = FALSE,
  contourSurfacePlotResponseDivision = 5,
  contourSurfacePlotType = "contourPlot",
  contourSurfacePlotVariables = list(types = list(), value = list()),
  covariates = list(types = list(), value = list()),
  dependentFactorial = list(types = list(), value = list()),
  dependentResponseSurface = list(types = list(), value = list()),
  designType = "factorialDesign",
  fixedFactorsFactorial = list(types = list(), value = list()),
  fixedFactorsResponseSurface = list(types = list(), value = list()),
  fourInOneResidualPlot = FALSE,
  highestOrder = TRUE,
  histogramBinWidthType = "sturges",
  histogramManualNumberOfBins = 30,
  modelTerms = list(optionKey = "components", types = list(), value = list()),
  normalEffectsPlot = FALSE,
  optimizationPlot = TRUE,
  optimizationPlotCustomParameterValues = list(optionKey = "variable", types = list(),
    value = list()),
  optimizationPlotCustomParameters = FALSE,
  optimizationPlotPredictionType = "response",
  optimizationSolutionTable = TRUE,
  order = 2,
  plotFitted = FALSE,
  plotHeight = 320,
  plotHist = FALSE,
  plotNorm = FALSE,
  plotPareto = FALSE,
  plotRunOrder = FALSE,
  plotWidth = 480,
  responseOptimizerManualBounds = FALSE,
  responseOptimizerManualTarget = FALSE,
  responsesResponseOptimizer = list(optionKey = "variable", types = list(), value =
    list()),
  rsmPredefinedModel = FALSE,
  rsmPredefinedTerms = "fullQuadratic",
  squaredTerms = list(types = list(), value = list()),
  squaredTermsCoded = FALSE,
  stepwiseMethodFactorial = "enter",
  stepwiseMethodResponseSurface = "enter",
  sumOfSquaresType = "type3",
  surfacePlotHorizontalRotation = 330,
  surfacePlotVerticalRotation = 20,
  tableAlias = TRUE,
  tableEquation = TRUE
)
}
\arguments{
\item{stepwiseMethodFactorial, }{Specify the order in which the predictors are entered into the model. A block of one or more predictors represents one step in the hierarchy. Note that the present release does not allow for more than one block. The backward and forward method are based on model AIC.
\itemize{
\item \code{"enter"} (default) : All predictors are entered into the model simultaneously.
\item \code{"backward"}: Starting with the full model, predictors are removed sequentially based on AIC.
\item \code{"forward"}: Starting with the intercept-only model, predictors are entered sequentially based on AIC.
\item \code{"both"}: Starting with the intercept-only model, predictors are entered or removed sequentially based on AIC, combining both forward addition and backward elimination at each step.
}}

\item{stepwiseMethodResponseSurface, }{Specify the order in which the predictors are entered into the model. A block of one or more predictors represents one step in the hierarchy. Note that the present release does not allow for more than one block.
\itemize{
\item \code{"enter"} (default) : All predictors are entered into the model simultaneously.
\item \code{"backward"}: Starting with the full model, predictors are removed sequentially based on AIC.
\item \code{"forward"}: Starting with the intercept-only model, predictors are entered sequentially based on AIC.
\item \code{"both"}: Starting with the intercept-only model, predictors are entered or removed sequentially based on AIC, combining both forward addition and backward elimination at each step.
}}
}
\description{
Analyse Design
}
