% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeFactorialWrapper.R
\name{doeFactorial}
\alias{doeFactorial}
\title{Create Factorial Worksheet}
\usage{
doeFactorial(
  data = NULL,
  version = "1",
  actualExporter = FALSE,
  blocks = "1",
  categoricalNoLevels = 2,
  categoricalVariables = list(list(levels = list("Row 0", "Row 1", "Row 2"), name =
    "data 1", values = list("A", "B", "C")), list(levels = list("Row 0", "Row 1",
    "Row 2"), name = "data 2", values = list("a", "a", "a")), list(levels = list("Row 0",
    "Row 1", "Row 2"), name = "data 3", values = list("b", "b", "b"))),
  centerpoints = 0,
  codedOutput = FALSE,
  displayDesign = TRUE,
  exportDesignFile = "",
  factorialDesignTypeSplitPlotNumberHardToChangeFactors = 1,
  factorialType = "factorialTypeDefault",
  factorialTypeSpecifyGenerators = "",
  numberOfCategorical = 3,
  plotHeight = 320,
  plotWidth = 480,
  replications = 1,
  runOrder = "runOrderRandom",
  seed = 1,
  selectedCol = -1,
  selectedDesign2 = list(list(levels = list("Row 0", "Row 1"), name = "data 1", values =
    list(4, 8)), list(levels = list("Row 0", "Row 1"), name = "data 2", values = list(0,
    0))),
  selectedRow = -1,
  setSeed = FALSE,
  showAliasStructure = FALSE
)
}
\description{
Create Factorial Worksheet
}
