% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeResponseSurfaceMethodologyWrapper.R
\name{doeResponseSurfaceMethodology}
\alias{doeResponseSurfaceMethodology}
\title{Create Response Surface Worksheet}
\usage{
doeResponseSurfaceMethodology(
  data = NULL,
  version = "1",
  actualExporter = FALSE,
  alphaType = "default",
  categoricalNoLevels = 2,
  categoricalVariables = list(list(levels = list(), name = "data 1", values = list()),
    list(levels = list(), name = "data 2", values = list()), list(levels = list(), name =
    "data 3", values = list())),
  centerPointType = "default",
  codedOutput = FALSE,
  continuousVariables = list(list(levels = list("Row 0", "Row 1"), name = "data 1",
    values = list("A", "B")), list(levels = list("Row 0", "Row 1"), name = "data 2",
    values = list(-1, -1)), list(levels = list("Row 0", "Row 1"), name = "data 3", values
    = list(1, 1))),
  customAlphaValue = 0,
  customAxialBlock = 0,
  customCubeBlock = 0,
  designType = "centralCompositeDesign",
  displayDesign = TRUE,
  exportDesignFile = "",
  numberOfCategorical = 0,
  numberOfContinuous = 2,
  plotHeight = 320,
  plotWidth = 480,
  replicates = 1,
  runOrder = "runOrderRandom",
  seed = 1,
  selectedCol = -1,
  selectedDesign2 = list(list(levels = list("Row 0", "Row 1"), name = "data 1", values =
    list(13, 14)), list(levels = list("Row 0", "Row 1"), name = "data 2", values =
    list(5, 6)), list(levels = list("Row 0", "Row 1"), name = "data 3", values = list(0,
    3)), list(levels = list("Row 0", "Row 1"), name = "data 4", values = list(0, 3)),
    list(levels = list("Row 0", "Row 1"), name = "data 5", values = list(1.414, 1.414))),
  selectedRow = -1,
  setSeed = FALSE
)
}
\description{
Create Response Surface Worksheet
}
