% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaAttributeWrapper.R
\name{msaAttribute}
\alias{msaAttribute}
\title{Attributes Agreement Analysis}
\usage{
msaAttribute(
  data = NULL,
  version = "1",
  cohensKappa = FALSE,
  dataFormat = "longFormat",
  fleissKappa = FALSE,
  kendallsTau = FALSE,
  measurementLongFormat = list(types = list(), value = ""),
  measurementsWideFormat = list(types = list(), value = list()),
  operatorLongFormat = list(types = list(), value = ""),
  operatorWideFormat = list(types = list(), value = ""),
  partLongFormat = list(types = list(), value = ""),
  partWideFormat = list(types = list(), value = ""),
  plotHeight = 320,
  plotWidth = 480,
  positiveReference = "",
  standardLongFormat = list(types = list(), value = ""),
  standardWideFormat = list(types = list(), value = "")
)
}
\description{
Attributes Agreement Analysis
}
