% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaGaugeLinearityWrapper.R
\name{msaGaugeLinearity}
\alias{msaGaugeLinearity}
\title{Type 4 Linearity Study}
\usage{
msaGaugeLinearity(
  data = NULL,
  version = "1",
  biasTable = TRUE,
  linearityAndBiasPlot = TRUE,
  linearityTable = TRUE,
  manualProcessVariation = FALSE,
  manualProcessVariationValue = 1,
  measurement = list(types = list(), value = ""),
  part = list(types = list(), value = ""),
  percentProcessVariationPlot = FALSE,
  plotHeight = 320,
  plotWidth = 480,
  standard = list(types = list(), value = "")
)
}
\description{
Type 4 Linearity Study
}
