% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaGaugeRRnonrepWrapper.R
\name{msaGaugeRRnonrep}
\alias{msaGaugeRRnonrep}
\title{Gauge r&R Study (Non-replicable Measurements)}
\usage{
msaGaugeRRnonrep(
  data = NULL,
  version = "1",
  anova = TRUE,
  dataFormat = "longFormat",
  historicalSdValue = 3,
  measurementLongFormat = list(types = list(), value = ""),
  measurementsWideFormat = list(types = list(), value = list()),
  operatorLongFormat = list(types = list(), value = ""),
  operatorMeasurementPlot = FALSE,
  operatorWideFormat = list(types = list(), value = ""),
  partLongFormat = list(types = list(), value = ""),
  partMeasurementPlot = FALSE,
  partMeasurementPlotAllValues = FALSE,
  partWideFormat = list(types = list(), value = ""),
  plotHeight = 320,
  plotWidth = 480,
  processVariationReference = "studySd",
  rChart = FALSE,
  report = FALSE,
  reportAverageChartByOperator = TRUE,
  reportCharacteristic = TRUE,
  reportCharacteristicText = "",
  reportDate = TRUE,
  reportDateText = "",
  reportGaugeName = TRUE,
  reportGaugeNameText = "",
  reportGaugeNumber = TRUE,
  reportGaugeNumberText = "",
  reportGaugeTable = TRUE,
  reportLocation = TRUE,
  reportLocationText = "",
  reportMeasurementsByOperatorPlot = TRUE,
  reportMetaData = TRUE,
  reportPartName = TRUE,
  reportPartNameText = "",
  reportPerformedBy = TRUE,
  reportPerformedByText = "",
  reportRChartByOperator = TRUE,
  reportTitle = TRUE,
  reportTitleText = "",
  reportTolerance = TRUE,
  reportToleranceText = "",
  reportTrafficLightChart = TRUE,
  reportVariationComponents = TRUE,
  rule1 = TRUE,
  rule2 = TRUE,
  rule2Value = 7,
  rule3 = TRUE,
  rule3Value = 7,
  rule4 = TRUE,
  rule4Value = 2,
  rule5 = TRUE,
  rule5Value = 15,
  rule6 = TRUE,
  rule6Value = 8,
  rule7 = FALSE,
  rule7Value = 4,
  rule8 = FALSE,
  rule8Value = 14,
  studyVarianceMultiplierType = "sd",
  studyVarianceMultiplierValue = 6,
  testSet = "jaspDefault",
  tolerance = FALSE,
  toleranceValue = 1,
  trafficLightChart = FALSE,
  varianceComponentsGraph = TRUE,
  xBarChart = FALSE
)
}
\description{
Gauge r&R Study (Non-replicable Measurements)
}
