% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaTestRetestWrapper.R
\name{msaTestRetest}
\alias{msaTestRetest}
\title{Test-retest (Range method)}
\usage{
msaTestRetest(
  data = NULL,
  version = "1",
  dataFormat = "longFormat",
  manualProcessSd = FALSE,
  manualProcessSdValue = 1,
  measurementLongFormat = list(types = list(), value = ""),
  measurementsWideFormat = list(types = list(), value = list()),
  operator = list(types = list(), value = ""),
  partLongFormat = list(types = list(), value = ""),
  partWideFormat = list(types = list(), value = ""),
  plotHeight = 320,
  plotWidth = 480,
  rChart = FALSE,
  repeatabilityAndReproducibilityTable = TRUE,
  runChartPart = FALSE,
  scatterPlotMeasurement = TRUE,
  scatterPlotMeasurementAllValues = FALSE,
  scatterPlotMeasurementFitLine = TRUE,
  tolerance = FALSE,
  toleranceValue = 1,
  trafficLightChart = FALSE
)
}
\description{
Test-retest (Range method)
}
