% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaType1GaugeWrapper.R
\name{msaType1Gauge}
\alias{msaType1Gauge}
\title{Type 1 Instrument Capability Study}
\usage{
msaType1Gauge(
  data = NULL,
  version = "1",
  biasTable = TRUE,
  histogram = FALSE,
  histogramBinBoundaryDirection = "left",
  histogramBinWidthType = "sturges",
  histogramManualNumberOfBins = 30,
  histogramMeanCi = TRUE,
  histogramMeanCiLevel = 0.95,
  histogramMeanLine = TRUE,
  histogramReferenceValueLine = TRUE,
  measurement = list(types = list(), value = ""),
  percentToleranceForCg = 20,
  plotHeight = 320,
  plotWidth = 480,
  referenceValue = 0,
  runChart = TRUE,
  runChartIndividualMeasurementDots = TRUE,
  runChartToleranceLimitLines = TRUE,
  studyVarianceMultiplier = "6",
  tTest = TRUE,
  tTestCiLevel = 0.95,
  toleranceRange = 1
)
}
\description{
Type 1 Instrument Capability Study
}
