% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilityOfDetectionWrapper.R
\name{probabilityOfDetection}
\alias{probabilityOfDetection}
\title{Probability of Detection}
\usage{
probabilityOfDetection(
  data = NULL,
  version = "1",
  covariate = list(types = list(), value = ""),
  detectionPlotCi = FALSE,
  detectionPlotCiLevel = 0.95,
  detectionPlotDataDisplay = TRUE,
  detectionPlotDataDisplayType = "rug",
  detectionPlotDataDisplayTypePointsJitter = TRUE,
  detectionPlotDensityDisplay = TRUE,
  horizontalAsymptotes = list(),
  linkFunction = "logit",
  logTransformedCovariate = FALSE,
  logarithmicXAxis = FALSE,
  modelFitTable = FALSE,
  outcome = list(types = list(), value = list()),
  plotHeight = 320,
  plotWidth = 480,
  verticalAsymptotes = list(),
  xAxisTicksType = "dataBased"
)
}
\description{
Probability of Detection
}
