% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareEventChartsWrapper.R
\name{rareEventCharts}
\alias{rareEventCharts}
\title{Rare Event Charts}
\usage{
rareEventCharts(
  data = NULL,
  version = "1",
  dataType = "dataTypeDates",
  dataTypeDatesFormatDate = "dm",
  dataTypeDatesFormatTime = "HM",
  dataTypeDatesStructure = "dateTime",
  dataTypeIntervalTimeFormat = "HM",
  dataTypeIntervalType = "opportunities",
  gChart = TRUE,
  gChartHistoricalProportion = 0.5,
  gChartProportionSource = "data",
  plotHeight = 320,
  plotWidth = 480,
  report = FALSE,
  reportChartName = TRUE,
  reportChartNameText = "",
  reportDate = TRUE,
  reportDateText = "",
  reportFootnote = TRUE,
  reportFootnoteText = "",
  reportLocation = TRUE,
  reportLocationText = "",
  reportMeasurementName = TRUE,
  reportMeasurementNameText = "",
  reportMetaData = TRUE,
  reportPerformedBy = TRUE,
  reportPerformedByText = "",
  reportPrintDate = TRUE,
  reportPrintDateText = "",
  reportSubtitle = TRUE,
  reportSubtitleText = "",
  reportTitle = TRUE,
  reportTitleText = "",
  rule1 = TRUE,
  rule2 = FALSE,
  rule2Value = 9,
  rule3 = FALSE,
  rule3Value = 6,
  rule8 = FALSE,
  rule8Value = 14,
  rule9 = FALSE,
  rule9Value = 3,
  stage = list(types = list(), value = ""),
  tChart = FALSE,
  tChartDistribution = "weibull",
  tChartDistributionParameterSource = "data",
  tChartHistoricalParametersScale = 2,
  tChartHistoricalParametersWeibullShape = 2,
  testSet = "jaspDefault",
  variable = list(types = list(), value = "")
)
}
\description{
Rare Event Charts
}
