% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeWeightedChartsWrapper.R
\name{timeWeightedCharts}
\alias{timeWeightedCharts}
\title{Time Weighted Charts}
\usage{
timeWeightedCharts(
  data = NULL,
  version = "1",
  axisLabels = list(types = list(), value = ""),
  cumulativeSumChart = TRUE,
  cumulativeSumChartAverageMovingRangeLength = 2,
  cumulativeSumChartNumberSd = 4,
  cumulativeSumChartSdMethod = "averageMovingRange",
  cumulativeSumChartSdSource = "data",
  cumulativeSumChartSdValue = 3,
  cumulativeSumChartShiftSize = 0.5,
  cumulativeSumChartTarget = 0,
  dataFormat = "longFormat",
  exponentiallyWeightedMovingAverageChart = FALSE,
  exponentiallyWeightedMovingAverageChartLambda = 0.3,
  exponentiallyWeightedMovingAverageChartMovingRangeLength = 2,
  exponentiallyWeightedMovingAverageChartSdMethod = "averageMovingRange",
  exponentiallyWeightedMovingAverageChartSdSource = "data",
  exponentiallyWeightedMovingAverageChartSdValue = 3,
  exponentiallyWeightedMovingAverageChartSigmaControlLimits = 3,
  groupingVariableMethod = "newLabel",
  manualSubgroupSizeValue = 5,
  measurementLongFormat = list(types = list(), value = ""),
  measurementsWideFormat = list(types = list(), value = list()),
  plotHeight = 320,
  plotWidth = 480,
  report = FALSE,
  reportChartName = TRUE,
  reportChartNameText = "",
  reportDate = TRUE,
  reportDateText = "",
  reportFootnote = TRUE,
  reportFootnoteText = "",
  reportLocation = TRUE,
  reportLocationText = "",
  reportMeasurementName = TRUE,
  reportMeasurementNameText = "",
  reportMetaData = TRUE,
  reportPerformedBy = TRUE,
  reportPerformedByText = "",
  reportPrintDate = TRUE,
  reportPrintDateText = "",
  reportSubtitle = TRUE,
  reportSubtitleText = "",
  reportTitle = TRUE,
  reportTitleText = "",
  rule1 = TRUE,
  stagesLongFormat = list(types = list(), value = ""),
  stagesWideFormat = list(types = list(), value = ""),
  subgroup = list(types = list(), value = ""),
  subgroupSizeType = "individual"
)
}
\description{
Time Weighted Charts
}
