% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variablesChartsIndividualsWrapper.R
\name{variablesChartsIndividuals}
\alias{variablesChartsIndividuals}
\title{Variables Charts for Individuals}
\usage{
variablesChartsIndividuals(
  data = NULL,
  version = "1",
  autocorrelationPlot = FALSE,
  autocorrelationPlotCiLevel = 0.95,
  autocorrelationPlotLagsNumber = 25,
  axisLabels = list(types = list(), value = ""),
  controlLimitsNumberOfSigmas = 3,
  measurement = list(types = list(), value = ""),
  plotHeight = 320,
  plotWidth = 480,
  report = FALSE,
  reportAutocorrelationChart = FALSE,
  reportChartName = TRUE,
  reportChartNameText = "",
  reportDate = TRUE,
  reportDateText = "",
  reportFootnote = TRUE,
  reportFootnoteText = "",
  reportIMRChart = TRUE,
  reportLocation = TRUE,
  reportLocationText = "",
  reportMeasurementName = TRUE,
  reportMeasurementNameText = "",
  reportMetaData = TRUE,
  reportPerformedBy = TRUE,
  reportPerformedByText = "",
  reportPrintDate = TRUE,
  reportPrintDateText = "",
  reportSubtitle = TRUE,
  reportSubtitleText = "",
  reportTitle = TRUE,
  reportTitleText = "",
  rule1 = TRUE,
  rule2 = TRUE,
  rule2Value = 7,
  rule3 = TRUE,
  rule3Value = 7,
  rule4 = TRUE,
  rule4Value = 2,
  rule5 = TRUE,
  rule5Value = 15,
  rule6 = TRUE,
  rule6Value = 8,
  rule7 = FALSE,
  rule7Value = 4,
  rule8 = FALSE,
  rule8Value = 14,
  stage = list(types = list(), value = ""),
  testSet = "jaspDefault",
  xmrChart = TRUE,
  xmrChartMovingRangeLength = 2
)
}
\description{
Variables Charts for Individuals
}
