% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variablesChartsSubgroupsWrapper.R
\name{variablesChartsSubgroups}
\alias{variablesChartsSubgroups}
\title{Variables Charts for Subgroups}
\usage{
variablesChartsSubgroups(
  data = NULL,
  version = "1",
  axisLabels = list(types = list(), value = ""),
  chartType = "xBarAndS",
  controlLimitsNumberOfSigmas = 3,
  dataFormat = "longFormat",
  fixedSubgroupSizeValue = 5,
  groupingVariableMethod = "newLabel",
  knownParameters = FALSE,
  knownParametersMean = 0,
  knownParametersSd = 3,
  manualSubgroupSizeValue = 5,
  measurementLongFormat = list(types = list(), value = ""),
  measurementsWideFormat = list(types = list(), value = list()),
  plotHeight = 320,
  plotWidth = 480,
  report = FALSE,
  reportChartName = TRUE,
  reportChartNameText = "",
  reportDate = TRUE,
  reportDateText = "",
  reportFootnote = TRUE,
  reportFootnoteText = "",
  reportLocation = TRUE,
  reportLocationText = "",
  reportMeasurementName = TRUE,
  reportMeasurementNameText = "",
  reportMetaData = TRUE,
  reportPerformedBy = TRUE,
  reportPerformedByText = "",
  reportPrintDate = TRUE,
  reportPrintDateText = "",
  reportSubtitle = TRUE,
  reportSubtitleText = "",
  reportTitle = TRUE,
  reportTitleText = "",
  rule1 = TRUE,
  rule2 = TRUE,
  rule2Value = 7,
  rule3 = TRUE,
  rule3Value = 7,
  rule4 = TRUE,
  rule4Value = 2,
  rule5 = TRUE,
  rule5Value = 15,
  rule6 = TRUE,
  rule6Value = 8,
  rule7 = FALSE,
  rule7Value = 4,
  rule8 = FALSE,
  rule8Value = 14,
  stagesLongFormat = list(types = list(), value = ""),
  stagesWideFormat = list(types = list(), value = ""),
  subgroup = list(types = list(), value = ""),
  subgroupSizeType = "manual",
  subgroupSizeUnequal = "actualSizes",
  testSet = "jaspDefault",
  warningLimits = FALSE,
  xBarAndSUnbiasingConstant = TRUE
)
}
\description{
Variables Charts for Subgroups
}
